Refactor: 함수 분리 및 description 변수명 변경
